+++ projects/addon-editor/components/color-selector/color-selector.component.ts
@@ -10,10 +10,13 @@ import {DomSanitizer, SafeStyle} from '@angular/platform-browser';
 import {TUI_EDITOR_COLOR_SELECTOR_MODE_NAMES} from '@taiga-ui/addon-editor/tokens';
 import {TuiGradientDirection} from '@taiga-ui/addon-editor/types';
 import {tuiGetGradientData, tuiParseGradient} from '@taiga-ui/addon-editor/utils';
-// note: unable to write a reference to TuiReorderComponent
-// eslint-disable-next-line @taiga-ui/no-deep-imports
-import {tuiDefaultSort} from '@taiga-ui/addon-table/utils';
-import {tuiDefaultProp, tuiParseColor, tuiPure, tuiRequiredSetter} from '@taiga-ui/cdk';
+import {
+    tuiDefaultProp,
+    tuiDefaultSort,
+    tuiParseColor,
+    tuiPure,
+    tuiRequiredSetter,
+} from '@taiga-ui/cdk';
 import {TuiHostedDropdownComponent} from '@taiga-ui/core';
 
 const EMPTY_STOP: [number, number, number, number] = [0, 0, 0, 0];
@@ -191,7 +194,7 @@ export class TuiColorSelectorComponent {
 
     private getGradient(direction: TuiGradientDirection): string {
         return `linear-gradient(${direction}, ${[...this.stopsKeys]
-            .sort((a, b) => tuiDefaultSort(a, b))
+            .sort(tuiDefaultSort)
             .map(key => `rgba(${this.getStop(key).join(', ')}) ${key * 100}%`)
             .join(', ')})`;
     }
