+++ projects/demo/src/modules/tables/table/examples/4/index.ts
@@ -2,11 +2,12 @@ import {Component} from '@angular/core';
 import {FormControl} from '@angular/forms';
 import {changeDetection} from '@demo/emulate/change-detection';
 import {encapsulation} from '@demo/emulate/encapsulation';
-import {TuiComparator, tuiDefaultSort} from '@taiga-ui/addon-table';
+import {TuiComparator} from '@taiga-ui/addon-table';
 import {
     TUI_DEFAULT_MATCHER,
     tuiControlValue,
     TuiDay,
+    tuiDefaultSort,
     tuiIsFalsy,
     tuiIsPresent,
     tuiToInt,
